Replace Travis with Gitlab CI
Travis provides a decent CI experience, but it does not provide any
pipeline features similar to Gitlab. As a result, steps that only have
to run once (such as linting), are performed for every Python version.
The pipeline provided by Gitlab gives more control, and results in
a "cleaner" build system.

In addition, the CI now also uploads coverage reports to Codecov.io.