Make cycle detection deterministic
The `simple_cycle` functions from networkx can return the cycle in an
arbitrary order. This causes non-deterministic (and therefore
unreliable) results. This implementation keeps track of previously
encountered nodes in a list that is past to the recursive call. It
slightly modifies the behaviour, especially with a self-loop, but this
is reflected in the unit tests.