Replace the learn functions to use dedup output
Instead of directly learning the model from the models directory,
learning now happens on the output of 'dedup'. This separates the
deduplication from learning.

There are now also two learning functions: dynamic distinguishing graph
(this similar what was already in place) and adaptive distinguishing
graph (which uses the output from adg-finder). Both methods can be
specified on the CLI.