Rewrite how benchmark data is generated
Input generation now uses a pandas DataFrame. This makes certain
operations easier, such as combinations with mappings.

Another important thing, is that it the benchmarks no longer include
combinations of different weight functions, and input selectors that do
not use weights. This comparison can still be made in the visualization,
but do not have to be recomputed here, which saves quite some time.

It also changes how results are stored. Instead of an including
a complete entry for each iteration, the measurements are stored as
a list. The naming of the fields is also updated.

This change requires that the visualization code is also updated.